First pass at cleaning out 64-bit issues.
Touches a lot of code to remove long constants like "1L", so this patch is
large and ugly, but I think it makes all those Clamp() calls look nicer in
the long run.

Most of the game is 64-bit clean, since we can build without assembly code
now. I've marked the things that are obviously still wrong with STUBBED lines.

That being said: a 64-bit build can already run the demos mostly correctly,
so we're actually almost there!

There are a few obvious things that are obviously wrong, to be fixed.

+++ Sources/Engine/Base/Types.h
@@ -26,12 +26,13 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 #include <Engine/Base/Base.h>
 #include <Engine/Graphics/gl_types.h>
 
-typedef signed long  int    SLONG;
+// !!! FIXME: use stdint.h for this (and other things like INDEX, too)?
+typedef signed int    SLONG;
 typedef signed short int    SWORD;
 typedef signed char         SBYTE;
 typedef signed int          SINT;
 
-typedef unsigned long  int  ULONG;
+typedef unsigned int  ULONG;
 typedef unsigned short int  UWORD;
 typedef unsigned char       UBYTE;
 typedef unsigned int        UINT;
@@ -282,8 +283,8 @@ typedef unsigned int        UINT;
 #define MAX_UBYTE ((UBYTE)0xFF)
 
 typedef int BOOL;		        // this is for TRUE/FALSE
-typedef long int RESULT;		// for error codes
-typedef long int INDEX;     // for indexed values and quantities
+typedef int RESULT;		// for error codes
+typedef int INDEX;     // for indexed values and quantities
 
 #define FALSE 0
 #define TRUE  1
